+++ VolumeRendering/GPURenderer4.cu
@@ -35,7 +35,7 @@ __device__ float4 sample_color_texture(float3 pos) {
 }
 
 __global__ void render_ray(uchar4 dev_buffer[]) {
-	int2 pos = {blockIdx.x * blockDim.x + threadIdx.x, blockIdx.y * blockDim.y + threadIdx.y};
+	short2 pos = {blockIdx.x * blockDim.x + threadIdx.x, blockIdx.y * blockDim.y + threadIdx.y};
 	if ((pos.x >= raycaster.view.size_px.x) || (pos.y >= raycaster.view.size_px.y))	// ak su rozmery okna nedelitelne 16, spustaju sa prazdne thready
 		return;
